Fix CTabFolder top-right control layout
RAP is using very old implementation of the server-side CTabFolder
control. In order to be used in e4, the widget was completely rewriten
by the SWT team. To avoid some layouting problems of the top-right
CTabFolder control, add missing redraw call to recalculate it's bounds.
Increase CTabItem top/bottom padding a little in default theme to avoid
the top-right control cut-off.

435499: CTabFolder#setTopRight does not layout appropriately when
modified
https://bugs.eclipse.org/bugs/show_bug.cgi?id=435499

512943: [e4] Rendering of toolbar broken after switching parts
https://bugs.eclipse.org/bugs/show_bug.cgi?id=512943

Change-Id: I38c99d6bc7027bcf95279f19888981af6bd0d197