Upgrade RAP Runtime build to Eclipse Tycho 1.0.0
Upgrade Eclipse Tycho for building to version 1.0 following the old
tradition of using the latest published build for building which brings
it to the same level as the Eclipse Platform build for Oxygen.

The release notes do *not* list any incompatibilities that would affect
the RAP Tools build:

https://wiki.eclipse.org/Tycho/Release_Notes/0.26
https://wiki.eclipse.org/Tycho/Release_Notes/1.0.0

One advantage is the usage of checksums especially for external bundles.

In order to minimize the risk of this upgrade, keep the old default
settings and do not take the project specific compiler settings into
account. This setting should be revised at a later time.

Change-Id: I0c99dbd8db5e02c22e521a76f1ddbc0164ec4d40
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>