Fix client crash when disposing an already disposed control
When widget is re-parented and disposed in the same request only
"distroy" operation is rendered. If its old parent (currently set on the
client) is also disposed in the same request, the widget "destroy"
operation is executed on the client on already removed from object
registry widget. This leads to a crash in
MessageProcessor#_processDestroy.

Add a check for available "objectEntry" in
MessageProcessor#_processDestroy.
Ensure that the parent in RemoteAdapter is in sync with widget parent
after re-parenting.

517403: [e4] JS error occurs when close view
https://bugs.eclipse.org/bugs/show_bug.cgi?id=517403

Change-Id: If7d7ff1a752aef67fd447218dad1bcf8466215b5