Fix unwanted scrolling when cell editor is opened
Focusing a control inside ScrolledComposite (from the server-side)
scrolls it into the view automatically. If control bounds are set in the
same response the focus code (scrolling) is executed before the new
bounds are flushed into the DOM. As a result, the scrolling happens on
the old control bounds.

Delay focus() call in Display.js#setFocusControl to ensure that the
control bounds has been applied to the DOM before element is focused.

497098: [ScrolledComposite] setShowFocusedControl Table jumps
unexpectedly when clicked
https://bugs.eclipse.org/bugs/show_bug.cgi?id=497098


Change-Id: I6c20daf713bdac367a8c0f74665061cce2ba71fb