Adjust Grid DND feedbak position for inserting before or after
Inserting an item before or after a referenced item should be done in
the same hierarchy level with that item. Currently it's not possible to
predict (based on the feedback line) item position if it is dropped on
the first position of a grid branch - after parent item or before first
item in that branch.

Adjust feedback line x position (similar to SWT) to indicate the
hierachy level where the item will be dropped.

520744: Drag and drop: no visual distinction in tree between different
feedback types FEEDBACK_INSERT_BEFORE and FEEDBACK_INSERT_AFTER
https://bugs.eclipse.org/bugs/show_bug.cgi?id=520744

Change-Id: Id1b41e398724e725c46dad92164653baf26377ee