Restore GC state after resetting the clip
Currently GC state changes can be lost when modifying the clip, 
because restore() is used to get rid of the previous clip, 
which affects the whole HTML5 GC state.

Fixed by keeping track of the most recent state on the client 
by introducing the "resetClip" function, 
which re-applies the state restore() is called. 

Compared to earlier versions of this patch, 
tracking the state is now done entirely on the client 
for efficiency reasons. For some of my test-cases this reduces protocol
size to 1/3 by avoiding re-transmission of redundant state, 
however it also makes it hard to unit-test.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=521472

Change-Id: Ia249cddced200fcbfe44baa784dcadc31edd4bbe
Signed-off-by: Clemens Eisserer <linuxhippy@gmail.com>