Add support for size/time upload limit to FileDialog/FileUploadHandler
Add possibility to specify the max upload size in bytes and the max
upload duration in milliseconds. When the specified value is exceeded,
the upload is terminated and the corresponding exception is stored.
A list with all exception during the upload can be obtained by calling
FileDialog#getExceptions().

Introduce two new public classes:

- UploadSizeLimitExceededException
- UploadTimeLimitExceededException

Change-Id: Ic50f6edc01c359065832221df76da534cdc3b6fe