Fix missing Text focus gained/lost events on iOS in some cases
When "next"/"previous" buttons are used in the virtual keyboard under
iOS the focus is moved to the next/previous Text field, but focus events
are not fired.

These buttons behave like TAB/SHIFT TAB on the PC, but they don't fire
key events. Thus, moving the focus is not recognized by the client.

Change-Id: Ic7a3dc6b8a611c0f0a8f5ae9352c44bd955f40ce