Add support for markup in ControlDecorator text
The markup support is enabled like on other widgets by using
setData(RWT.MARKUP_ENABLED, Boolean.TRUE).
As ControlDecorator is internal widget, the setData approach can be
replaced with public method setMarkupEnabled, but it will not be
consistant with other widgets that support markup.

Add new RAP specific API to JFace ControlDecoration - enableMarkup().
As it's not possible to disable markup support, once it is enabled,
setMarkupEnabled(boolean) does not fit well here.

Clean up the original commented code as it makes the class hard to read.

Bug 529572: ControlDecorator markup support
Change-Id: I3793f3802c3c4ee109e8acc23a3b910d7dea2f04