Block scrolling in ScrolledComposite if Button is focused by mouse
When ScrolledComposite#showFocusedControl is set to true, all Scrollable
parents of the focused control are automatically scrolled in order to
show it on the screen. In case of nested ScrolledComposite this may lead
to widget (Button) move between mouse down and mouse up event when it is
focused by mouse. As a result, the Selection event of the Button is not
fired.

This is not a generic solution for the broblem, but only a workaround
for focused Button by mouse inside ScrolledComposite.

Bug 528481: Button click does not work in cascaded scrolled composite
when the outer composite jumps on focus out.

Change-Id: I3670b823c916cbe83e25a55121ce09024e6f65fa