Prevent userScroll event when List items are changed
userScroll event is ment to fire only by user interaction. When List
items are changed, the userScroll event is fired in same cases (scrolled
List) when changing scrollbars maximum in _updateScrollDimension. To
prevent this, the dedicated _internalChangeFlag flag is set to true
before changing the maximum value in _updateScrollDimension.

Bug: 532672 Javascript error when adding/removing items from a scrolling
List

Change-Id: I9d24f9c51f49625cef776b5139a7b1899ffc9d43