Clear widgets preserved state after setting properties from the client
Common widget/control properties are preserved on demand when the
corresponding public setter is used. Some widgets (like Shell) are still
using the public setter to apply these properties when the value is
changed on the client. As a result, the old server value is preserved
when client value is applied. This makes it imposible to reset the
property back to its original value.

Clear preserved on demend properties after reading the client changes.

Bug: 533412: Shell.setSize within SWT.Resize listener not working
https://bugs.eclipse.org/bugs/show_bug.cgi?id=533412

Change-Id: I45aec2edfebcb820d1a3a07c5ca9e730e67855b8