Modify checking non-printable key on Firefox.
In Firefox > 60, keypress event will not be fired if non-printable key.
This behavior is similar to other user agents. So we don't need to
distinguish in the checking pseudo key. Furthermore, we need to consume
these exceeded keypress event in Firefox of this new version.

This fixes Bug 534628 - Keyboard navigation broken in tree widget with
Firefox Nightly/Beta releases

Change-Id: I2200291f517caccc759d09199c4c8bdad9bba953
Signed-off-by: Mantaroh Yoshinaga <mantaroh@gmail.com>