Fix parent menu not closed on item execution
This is a regression introduced in 3.5 with commit
e3231d62527a4cb7a6ae73058fe59b66c722a01a. In case od cascade menus, we
should check the item's parent menu if pressed and not the current menu.

Bug 537964: Context menu with a submenu, keeps parent menu open on click
https://bugs.eclipse.org/bugs/show_bug.cgi?id=537964

Change-Id: I19ffe5044d2115d52716cce06f6e82ca9cee7f84