Fix ComboBoxCellEditor not visible in single editable column on tab
In case of single editable table column with ComboBoxCellEditor when TAB
is pressed, old and new cell editor is the same. ComboBoxCellEditor has
e key listener to deactivate the editor on TAB. Thus, the newly
activated ComboBoxCellEditor by TAB is immidiatelly deactivated because
of this listener. ComboBoxCellEditor is deactivated by focus out
listener and there is no need to listen for TAB character.

Change-Id: Ibdb347e8f590797c2867d7c149c4e46cac6fe29a