Allow to set the upload directory in FileDialog/DiskFileUploadReceiver
Allow developers to specify the directory where file upload stores the
files to.

When the directory is set and it is a directory, the reciever uses it.
When the directory does not yet exist, the reciever creates it.
When the directory is a file, not a directory, an IOException is thrown.
When no directory is set, the temp directory will be used, the same way
it worked before this change.

Bug 434180: [FileUpload] Create API to configure temporary directory and
size threshold
https://bugs.eclipse.org/bugs/show_bug.cgi?id=434180

Change-Id: I2424364280b20a91e561e0c508904573ac24da68
Signed-off-by: Moritz Aleithe <moritz.aleithe@yahoo.de>