Fix shell not restored after it's maximized and then minimized
If a shell is minimized after it was maximised before, setting the mode
to "maximised" do nothing as the shell already has "maximised" state
set. In this case "_restoreFromMinimized" function must be called
instead of "_maximize" in Windows.js#_applyMode.

Bug 546558: Shell  is not coming to top after maximizing and minimizing
https://bugs.eclipse.org/bugs/show_bug.cgi?id=546558

Change-Id: I0f536038db2a7d81f3225dcb68b2e86aca3e7576