Don't allow vertical scrollbar max value to exceed visible items count
When alternative scrolling is enabled on the client-side Grid, the
vertical scrollbar value is adjusted in Grid#_updateScrollHeight. If
there is only one partly visible item, Grid#_getLastPageRowCount returns
zero. As a result the scrollbar maximum value is set to a value bigger
with 1 than visible items.

Bug 545437: Enabling AutoHeight for GridVierwer leads to
IllegalArgumentException
https://bugs.eclipse.org/bugs/show_bug.cgi?id=545437

Change-Id: Id420afe45d49cbf5d4feb95306aab72ffaf55e0d