Add possibility to abort the upload process from the server
If the ClientFileUploader client service is used it is not possible to
abort the ongoing upload process in contrast to the FileUpload widget.
When the FileUpload widget is disposed, the upload request is aborted by
the client.

Introduce new API ClientFileUploader#abort(String) to let the server
instruct the client to abort the upload request with the given id. The
id is generated when the "submit" method is called.

Change-Id: I1b849493798f065a3cbd83388d6d660e9f483d29