Update build configuration for RAP Runtime 3.11.0 M1
- Use Eclipse Platform build S-4.14M1-201910091800
- Use Eclipse Orbit S20191007144714
- Use Eclipse Jetty 9.4.21.v20190926

Disable all cluster tests in SessionFailover and in SessionSerialization
as some tests are failing sporadically, others all the time after the
upgrade of Eclipse Platform and Jetty. At a later time we have to decide
if we invest the time to stabilise the test, or if we completely remove
them.

Also remove org.eclipse.equinox.util from the RAP Basic E4 Equinox
Target feature as it is not shipped by the Eclipse Equinox team any
more, and there are no references in RAP itself. See bug 533801
https://bugs.eclipse.org/bugs/show_bug.cgi?id=533801 for the background
information.

Change-Id: I32fcf6ad0af0d94c9b48f4eb07a03f534623c824
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>