Prevent JS crash when navigating in drop-down calendar
Navigating (change month/year with navigation buttons) in the drop-down
calendar triggers "changeDate" events. When min/max date is set on the
widget it's possible to fire an event with a date which is out of range.

Apply only dates which are in the chosen range (between min and max).

Bug 552137: [DateTime] RAP crashes in the browser when using a DropDown
DateTime in a certain way
https://bugs.eclipse.org/bugs/show_bug.cgi?id=552137

Change-Id: I06a87971afa92547f8484ada9a4fec050252c68e