Remove RWT cluster tests from Tycho build
Together with the Eclipse platform update in commit edce2a7, the RWT
cluster tests had to be disabled because of incompatibilities with the
new platform and the new Jetty execution environment in the Tycho build.
Because the cluster tests are still operational from a local IDE, it
makes sense to just omit them in the server-side builds, but allow
running them locally at the same time.

Remove the previously added @Ignore tags from the RWT cluster tests.

Remove the following test bundles and artifacts from the Tycho build and
disable their test run.
- org.eclipse.rap.rwt.cluster.testfixture
- org.eclipse.rap.rwt.cluster.testfixture.test
- org.eclipse.rap.rwt.cluster.test

Change-Id: I758b8bf5e2cfdcea05925c0eecf0020b59c685a5
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>