Fix a regression related to ToolBar focus
With commit 0fb96a1bfe8d922b17c0396b01dca4d98f1d65ff we made ToolBar non
focusable by mouse. This is not completelly correct. In SWT the ToolBar
is only focusable if it is created with SWT.FLAT style flag.

Render NO_FOCUS style flag on the client and switch focus behaviour
based on it.

Bug 570167: selecting an action inside a view does not give focus to the
view
https://bugs.eclipse.org/bugs/show_bug.cgi?id=570167

Change-Id: Ib37c9f5c2b2c58e233c6dbc26dad13d0acc41824