Make RAP test suite compatible with Java 11 under Eclipse IDE
Since Java 9 the classloader is no longer a URLClassLoader. This leads
to two problems:
1. Direct casting to URLClassLoader failed
2. getURLs/addUrl methods are not available

To fix the issue the classpath is parsed directly from the system
property in TestCollector. As the code in ThemesTestUtil to dynamically
add folder to the classpath does not work anymore in Java 11 a static
classpath entry to the "org.eclipse.rap.design.example" bundle is added.

Change-Id: Ice8baf586dac1425a934e40656a2bc17334c002e