Update to Jetty 10.0.3 for RAP Runtime 3.17.0 M3
This minimised change updates Eclipse Jetty from 9.4.x to 10.0.3 by
courtesy of the WebTools project. It removes depependencies that are not
available any more, and adds empty stubs for missing methods in the new
ServletContextWrapper. Commits that were required for temporarily
reactivate the old Jetty version in older milestones are reverted.

Also updates Eclipse Platform to 4.20 M3 build I20210519-1800.

This reverts commit 334f03bef1da2290319dac7e3cfaab7796296c04.
This reverts commit f1041cb5069ed4726f3d9a958be0dc8c4179cdbb.

Change-Id: I9a6b2f8e1cf16364bc863afa23d967e3bb77715a
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>