Remove javax.servlet spec from exemplary war file builds
The servlet spec is required during build time only, and used to be
contributed by the javax.servlet bundle. With the upgrade of the spec
version, these Java packages are not contributed by the new
org.eclipse.jetty.servlet-api bundle which also needs to be excluded
from the .war files.

Create additional exclude rules for all three example products.

Change-Id: I2907c74cecb1c2784f14c8ed1cbbf6d94e5963cf
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>