Resolve Jetty 10 dependency to SLF4J in Equinox features
Jetty 10 bundles that are delivered as part of the Equinox features
which provide a ready to run minimal set of OSGi bundles for running RAP
applications require the availability of SLF4J.

Add the org.slf4j bundles from Orbit in version 1.7 to both features.
This bundle version is the same as provided in current Eclipse Platform
builds, and the lowest version accepted by Jetty.

Change-Id: I6964bfc77f11d333eecd92f166890b6c5b076c8f
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>