Ensure spinner min/max value is correctly adjusted when date is set
If the current date in the widget has spinner (day from February) max
value set to 28 and new day value is set to 31, the spinner does a
roll-up. The spinner max value must be in sync with the currently set
month.

Bug 578732: Wrong UI representation for date after editing text field in
DateTime widget
https://bugs.eclipse.org/bugs/show_bug.cgi?id=578732

Change-Id: Ibc6128a83e02444ed5086e603570a8f1dd7b5310