Upgrade Apache commons-fileupload to version 1.4.0
This change also includes Apache `commons-io` in version 2.11.0.

Update lower version boundaries of Apache `commons-fileupload` and
`commons-io` to the version of the updated bundles in order to ensure
that the correct versions are selected at runtime.

The new bundles are consumed from Eclipse Platform which retrieves them
from Maven Central. As part of this process, the bundle symbolic name of
the bundles changed:

- `org.apache.commons.fileupload` to `org.apache.commons.commons-fileupload`
- `org.apache.commons.io` to `org.apache.commons.commons-io`

Update features, demo launch configurations, and demo product
definitions to the new name.

This fixes #41.