Add mode to make text-size-store session scoped
Setting org.eclipse.rap.rwt.textSizeStoreSessionScoped=true
allows to trade performance for correctness in environments
where clients return inconsistent text measurements.
By caching the text sizes in the UISession instead of the app context,
measurements only have to be consistent within the same client.
The downside of this approach is each client has to re-compute
each text the first time it is measured.