Fix encoding problems with resource bundles encoded in UTF-8 (#77)
In Java versions 8 and lower the default encoding was ISO8859-1 and
ResourceBundle.getBundle(...) returned a bundle with ISO8859-1 encoded
messages. As a workaround for UTF-8 encoded messages we have
Utf8ResourceBundle class. This class is no longer needed when using Java
8+ where the default encoding is UTF-8.

Make the two methods NLS.getISO8859_1Encoded(...) and
NLS.getUTF8Encoded(...) deprecated and introduce a new NLS.get(...).