Automatically include all dependencies in output p2 repositories
Enable the includeAllDependencies switch when assembling the final
distribution p2 repositories. This ensures that all (transitive)
dependencies are included and that the resulting p2 repository is
self-contained.

In particular this change leads to the inclusion of the SLF4J NOP
implementation bundle into the p2 repository unless RAP or any of the
upstream projects decide to provide a different implementation. We do
not include it in any of our features to make it more unlikely that it
gets mirrored into other p2 repositories, or used in any real products.