Add possibility to get startup parameters from initial POST request
The initial RAP request POST or GET does not initiate UI session. There
is a client service StartupParameters to extract the startup  parameters
from URL query string and send them to the server with the first UI
request. With the initial POST request this is not working.

In order to fix it, get the startup parameters from the initial GET/POST
request on the server and send them back to client-side Display object
during the creation in StartupJson#appendCreateDisplay.

Fix #58