Upgrade to new org.mozilla.rhino version provided by Orbit
Orbit is migrating from hand-crafted bundles to pristine versions from
Maven central. Part of this is the bundle symbolic name change from
org.mozilla.javascript to org.mozilla.rhino.

Because this includes an upgrade of the version, the reference in the
RWT build needs to be in sync again, otherwise the build breaks.