Provide workaround for E4 javax.servlet dependencies
The bundle org.eclipse.e4.emf.xpath has a dependency to the external
bundle org.apache.commomns.jxpath which is not developed any more. This
bundle itself has dependencies to various Java packages from the
javax.servlet.* namespace.

As a short term solution, provide the javax.servlet API by adding the
Jakarta servlet 4.0 bundle in addition to the 6.1 version temporarily to
the E4 Equinox feature in order to satisfy the runtime only constraint.