Fix file upload in Tomcat 11
With commit
https://github.com/apache/tomcat/commit/3f8a229be8cbdd99e718b1b72b5f9c0ea1421c23
using request#getParameter in Tomcat 11 will parse the request body. As
a result the file upload is broken as the request body is already
processed in RWTServlet by the getting connection/service handler id.

The fix is to manually parse the query string parameters for multipart
requests and not using request#getParameter.