Provide org.mozilla.rhino to clientbuilder from old Orbit repo
The current Orbit repository 2025-03 does not contain the old
org.mozilla.rhino bundle any more which is required by the RAP
ClientBuilder. Update to newer replacements would require a large change
to new APIs.

Because the org.mozilla.rhino is only used during build time, and not
distributed nor published by the RAP project, it seems safe to rely on
the old Orbit repository locally in the clientbuilder project until
there is a better fix.