Exclude `org.eclipse.osgi.util.source` from target features
The RAP Runtime e3 and e4 target features used to include the source
bundle `org.eclipse.osgi.util.source` which went missing in Eclipse
Platform build 4.36 RC1. This source bundle is not required to build and
to run RAP applications, but may be useful in the IDE.

Remove it from the target features that are responsible for assembling
all bundles including sources for development. It may be added at a
later time, once it appears in the upstream repository again.