Add ClientCanvas widget and tests
A ClientCanvas can be used like a SWT Canvas with the difference that a
client can draw on it's area, too.

Note: The client scripting "redraw" method (CanvasHandler.js#redraw)
does not clear the canvas properties anymore. It fires only client-side
"Paint" event. This client-side implementation is currently in sync with
the server-side GC one.